+++ src/main/java/kz/halykacademy/bookstore/web/books/Book.java
@@ -1,10 +1,8 @@
 package kz.halykacademy.bookstore.web.books;
 
-import kz.halykacademy.bookstore.web.genres.Genre;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
 
@@ -21,5 +19,5 @@ public class Book {
     private int numberOfPages;
     private int releaseYear;
     private List<String> genreList;
-    private double bookQuantity;
+    private int bookQuantity;
 }
